+++ packages/moonstone/ToggleButton/ToggleButton.js
@@ -10,7 +10,6 @@
  */
 
 import kind from '@enact/core/kind';
-import deprecate from '@enact/core/internal/deprecate';
 import React from 'react';
 import PropTypes from 'prop-types';
 import Pure from '@enact/ui/internal/Pure';
@@ -20,19 +19,6 @@ import Button from '../Button';
 import Skinnable from '../Skinnable';
 
 import css from './ToggleButton.module.less';
-
-const deprecateSmall = deprecate((small) => small ? 'small' : 'large',  {
-	name: 'moonstone/ToggleButton.ToggleButtonBase#small',
-	replacedBy: 'the `size` prop',
-	message: 'Use `size="small"` instead.',
-	since: '2.6.0',
-	until: '3.0.0'
-});
-
-function getSize (size, small) {
-	small = typeof small !== 'undefined' ? deprecateSmall(small) : 'large';
-	return size || small;
-}
 
 /**
  * A stateless [Button]{@link moonstone/Button.Button} that can be toggled by changing its
@@ -120,15 +106,6 @@ const ToggleButtonBase = kind({
 		size: PropTypes.string,
 
 		/**
-		 * Reduces the size of the button.
-		 *
-		 * @type {Boolean}
-		 * @deprecated replaced by prop `size='small'`
-		 * @public
-		 */
-		small: PropTypes.bool,
-
-		/**
 		 * Button text displayed in the 'off' state.
 		 *
 		 * If not specified, `children` will be used for 'off' button text.
@@ -165,7 +142,7 @@ const ToggleButtonBase = kind({
 	},
 
 	computed: {
-		className: ({selected, size, small, styler}) => styler.append({selected}, getSize(size, small)),
+		className: ({selected, size, styler}) => styler.append({selected}, size),
 		children: ({children, selected, toggleOnLabel, toggleOffLabel}) => {
 			let c = children;
 			if (selected && toggleOnLabel) {
@@ -180,7 +157,6 @@ const ToggleButtonBase = kind({
 	render: ({selected, ...rest}) => {
 		delete rest.toggleOffLabel;
 		delete rest.toggleOnLabel;
-		delete rest.small;
 
 		return (
 			<Button